Move visualize_process_tree into ProcessTree class and fix imports
