+++ practical/ProcessMining/group2/inductive_miner/src/process_tree.py
@@ -1,10 +1,24 @@
 import copy
 import itertools
+import os
+import sys
+SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(os.path.dirname(SCRIPT_DIR))
 
-from practical.ProcessMining.group2.inductive_miner.src.graph import *
-from practical.ProcessMining.group2.inductive_miner.src.event_log import *
-from practical.ProcessMining.group2.inductive_miner.src.directly_follows_graph import *
+from graphviz import Digraph
+from enum import Enum
 
+from src.graph import *
+from src.event_log import *
+from src.directly_follows_graph import *
+
+
+class Relations(Enum):
+    LOOP = 'O'# LOOP = 'loop'   # 
+    SEQUENCE =  '->'   #seq =  '->'
+    CHOICE = 'X'   #choice = 'X'
+    TAU = 'tau'   # tau = 'tau'
+    PARALLEL = '||'  #parallel = '||'
 
 class ProcessTree:
     def __init__(self, event_log: EventLog) -> None:
@@ -397,10 +411,10 @@ class ProcessTree:
         dfg.construct_dfg()
 
         cut_methods = [
-            (self.find_exclusive_choice_cut, self.exclusive_choice_split, 'X'),
-            (self.find_sequence_cut, self.sequence_split, '->'),
-            (self.find_parallel_cut, self.parallel_split, '||'),
-            (self.find_loop_cut, self.loop_split, 'O')
+            (self.find_exclusive_choice_cut, self.exclusive_choice_split, Relations.CHOICE.value),
+            (self.find_sequence_cut, self.sequence_split, Relations.SEQUENCE.value),
+            (self.find_parallel_cut, self.parallel_split, Relations.PARALLEL.value),
+            (self.find_loop_cut, self.loop_split, Relations.LOOP.value)
         ]
 
         for find_cut, process_split, operator in cut_methods:
@@ -421,6 +435,58 @@ class ProcessTree:
         # Fallthrough case
         return 'O', ['tau'] + dfg.get_all_nodes()
     
+    def visualize_process_tree(self, tree):
+        # Initialize a Graphviz Digraph object
+        dot = Digraph()
+        
+        # Recursive function to add nodes and edges to the graph
+        # node_id to keep track of hierarchy level of the tree
+        def add_nodes_edges(tree, parent=None, node_id=0):
+            if isinstance(tree, tuple):
+                node_label = tree[0]
+                children = tree[1]
+            else:
+                node_label = tree
+                children = []
+            
+            current_id = str(node_id)
+          
+            match node_label: 
+                case Relations.LOOP.value:
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/loop_arrow.png')
+                    dot.node(current_id, image=image_path, shape="ellipse", fixedsize="true", label="", width="0.6", height="0.7", penwidth = "2")  
+                case Relations.PARALLEL.value: 
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/and.png')  
+                    dot.node(current_id, image=image_path, shape="ellipse", label="", fixedsize="true", width="0.6", height="0.7", penwidth = "2")
+                case Relations.SEQUENCE.value:
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/seq_arrow.png') 
+                    dot.node(current_id, image=image_path, shape="ellipse", label="", fixedsize="true", width="0.6", height="0.7", penwidth = "2")
+                case Relations.CHOICE.value:
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/x.png') 
+                    dot.node(current_id, image=image_path, shape="ellipse", label="", fixedsize="true", width="0.6", height="0.7", penwidth = "2")
+                case Relations.TAU.value:
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/tau.png')
+                    dot.node(current_id, image=image_path, shape="square", label="", fixedsize="true", width="0.6", height="0.7", penwidth = "2",imagescale = "true")
+                case None:
+                    image_path = os.path.join(os.path.dirname(__file__), '../data/images/bomb.png')
+                    dot.node(current_id, image=image_path, imagescale = "true", shape="square", label="", fixedsize="true", width="0.6", height="0.7", penwidth = "2",fontname="Arial", fontcolor="black")
+                case _:   
+                    dot.node(current_id, node_label,  shape="square", fixedsize="true", width="0.6", height="0.7",  fontsize="30", fontname="Arial", fontcolor="black" ) #shape="circle"
+             
+         
+            if parent is not None:
+                dot.edge(parent, current_id)
+            
+            for i, child in enumerate(children):
+                add_nodes_edges(child, current_id, node_id * 10 + i + 1)
+        
+        # Start the recursion with the root of the tree
+        add_nodes_edges(tree)
+        
+      
+      
+        dot.render('tree', format='png', view=True)
+    
     def __str__(self) -> str:
         """
         Return the string representation of the process tree.
@@ -428,11 +494,11 @@ class ProcessTree:
         :return: String representation of the process tree.
         """
         operator_map = {
-            'O': '↺',
-            'X': 'x',
-            '->': '➜',
-            '||': '∧',
-            'tau': '𝝉'
+            Relations.LOOP.value: '↺',
+            Relations.CHOICE.value: 'x',
+            Relations.SEQUENCE.value: '➜',
+            Relations.PARALLEL.value: '∧',
+            Relations.TAU.value: '𝝉'
         }
         tree = self.construct_process_tree()
